+++ penman/model.py
@@ -17,7 +17,6 @@ from penman.types import (
     Target,
     BasicTriple
 )
-from penman.tree import Branch
 from penman.graph import CONCEPT_ROLE
 
 
@@ -278,13 +277,12 @@ class Model(object):
         raise ModelError(f'{source_role!r} and {target_role!r} '
                          f'are not valid roles to dereify {concept!r}')
 
-    def original_order(self, branch: Branch):
-        """Branch sorting key that does not change the order."""
+    def original_order(self, role: Role):
+        """Role sorting key that does not change the order."""
         return True
 
-    def canonical_order(self, branch: Branch):
-        """Branch sorting key that finds a canonical order."""
-        role, _ = branch
+    def canonical_order(self, role: Role):
+        """Role sorting key that finds a canonical order."""
         m = re.match(r'(.*\D)(\d+)$', role)
         if m:
             rolename = m.group(1)
@@ -293,6 +291,6 @@ class Model(object):
             rolename, roleno = role, 0
         return (self.is_role_inverted(role), rolename, roleno)
 
-    def random_order(self, branch: Branch):
-        """Branch sorting key that randomizes the order."""
+    def random_order(self, role: Role):
+        """Role sorting key that randomizes the order."""
         return random.random()