+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## [Unreleased][unreleased]
 
+(no unreleased changes yet)
+
+
+## [v0.10.0][]
+
+**Release date: 2020-01-16**
+
+Add triple sorting to the `reconfigure()` function and make it
+available at the command line.
+
 ### Added
 
 * `penman.layout.reconfigure()` now has a `key` parameter ([#52][])
@@ -15,7 +25,6 @@
   `Branch` argument ([#53][])
 * `penman.model.Model.random_order()` takes a `Role` instead of a
   `Branch` argument ([#53][])
-
 
 
 ## [v0.9.1][]
@@ -553,6 +562,7 @@ First release with very basic functionality.
 [v0.8.0]: ../../releases/tag/v0.8.0
 [v0.9.0]: ../../releases/tag/v0.9.0
 [v0.9.1]: ../../releases/tag/v0.9.1
+[v0.10.0]: ../../releases/tag/v0.10.0
 [README]: README.md
 
 [#4]: https://github.com/goodmami/penman/issues/4